requirements.txt
requests
langchain
langchain-openai
pymupdf
python-dotenv
tqdm
tiktoken
PyPDF2
pypdf
